wip2

+++ CoronaDVH/Models/DvhData.cs
@@ -6,17 +6,17 @@
         /// Discrete DVH data points (not cumulative)
         /// </summary>
         public List<DvhPoint> Points { get; set; } = new List<DvhPoint>();
-        public float MaxDose { get; set; }
-        public float MinDose { get; set; }
-        public float MeanDose { get; set; }
-        public float Volume { get; internal set; }
+        public double MaxDose { get; set; }
+        public double MinDose { get; set; }
+        public double MeanDose { get; set; }
+        public double Volume { get; internal set; }
 
         /// <summary>
         /// Samples the dose at a given volume for a cumulative DVH.
         /// </summary>
         /// <param name="sampleVolume">volume in the same units as the points</param>
         /// <returns>dose in the same units as the points</returns>
-        public float GetDoseAtVolume(float sampleVolume)
+        public double GetDoseAtVolume(float sampleVolume)
         {
             // Safety check: if there are no points, return 0 (or throw an exception).
             if (!Points.Any())
@@ -27,8 +27,8 @@
 
             // Compute cumulative volumes for the sorted list.
             // cumVolumes[i] is the cumulative volume from index 0 up to i.
-            float cumulativeVolume = 0f;
-            var cumList = new List<(float Dose, float CumVolume)>(sortedPoints.Count);
+            double cumulativeVolume = 0f;
+            var cumList = new List<(double Dose, double CumVolume)>(sortedPoints.Count);
             foreach (var pt in sortedPoints)
             {
                 cumulativeVolume += pt.Volume;
@@ -57,7 +57,7 @@
                     var (doseLow, volLow) = cumList[i];
 
                     // Calculate the fraction between volHigh and volLow that gives us sampleVolume.
-                    float fraction = (sampleVolume - volHigh) / (volLow - volHigh);
+                    double fraction = (sampleVolume - volHigh) / (volLow - volHigh);
 
                     // Since dose decreases as cumulative volume increases, we linearly interpolate.
                     // dose = doseHigh + fraction*(doseLow - doseHigh)
@@ -74,7 +74,7 @@
         /// </summary>
         /// <param name="dose">dose in the same units as the points</param>
         /// <returns>volume in the same units as the points</returns>
-        public float GetVolumeAtDose(float dose)
+        public double GetVolumeAtDose(float dose)
         {
             // Sum all volumes for which the dose is at least the specified threshold.
             return Points.Where(p => p.Dose >= dose).Sum(p => p.Volume);
